src/stack.py
class Node:
    """Класс для узла стека"""

    def __init__(self, data, next_node=None):
        """
        Конструктор класса Node

        :param data: данные, которые будут храниться в узле
        """
        self.data = data
        self.next_node = next_node


class Stack:
    """Класс для стека"""

    def __init__(self):
        """Конструктор класса Stack"""
        self.top = None


    def __iter__(self):
        current = self.top
        while current is not None:
            yield current.data
            current = current.next_node


    def __str__(self):
        return "Stack is empty" if self.top is None else 'Stack: ' + ', '.join([str(item) for item in self])


    def push(self, data):
        """
        Метод для добавления элемента на вершину стека

        :param data: данные, которые будут добавлены на вершину стека
        """
        new_node = Node(data)
        if self.top is None:
            self.top = new_node
        else:
            new_node.next_node = self.top
            self.top = new_node

    def pop(self):
        """
        Метод для удаления элемента с вершины стека и его возвращения

        :return: данные удаленного элемента
        """
        if self.top is None:
            raise Exception('Stack is empty')
        data = self.top.data
        self.top = self.top.next_node
        return data

